Relax checksum-related timeouts
Currently we only allow for 120 seconds (2 minutes) to calculate the
checksum of a volume dump. For sufficiently large volume dumps (dozens
of gigabytes), this may not be enough time, resulting in ALERT_EXPIRED
alerts for such volumes in the "Processing finished dump file" or
"Checking retrieved dump file" stages.

In a future commit, we should report progress as we calculate the
checksum, instead of just having a single timeout for the entire file.
For now, though, just raise the timeout to 1800 seconds (30 minutes),
so we're less likely to encounter this error.