Redesign checksum-related timeouts and reports
Currently we have a timeout of 1800 seconds (30 minutes) to
calculate the checksum of a volume dump. On average, this timeout
should be enough in most of the cases. However, a single timeout
for the entire file is not the best approach for a couple of reasons.
First of all, the user does not get any report about the current
progress of this state. As a result, he has no idea if CellCC is stuck
or making some progress. Secondly, this timeout might still not be
enough for the whole file.

In order to fix the problems reported above, a new function was
designed to calculate the checksum of the whole dump file. This
function breaks the file into smaller chunks, computes the checksum
of these smaller pieces, and add them all in separated steps. Also,
this function reports the progress of this state minute by minute.
Last, but not least, the timeout of the state in question is updated
every single minute.