Improve performance avoiding unnecessary vos dumps
Currently, every single job is completely independent of each other. As
a result, when a volume is being synchronized between several destination
cells, the same volume is dumped multiple times. In order to improve the
performance, this patch introduces a set of changes aiming to avoid
unnecessary time consuming dumps. To make it possible, the jobs
requesting the same volume with the same timestamp are grouped into a
hash table and only one dump file is generated per entry. The dump file
for a specific entry is assigned to the first job of the list and the
remaining jobs receive a copy of that dump file.

[adeason@sinenomine.net: reworked incremental calculations and
%lu_table population.]