DB: Allow find_jobs to accept arrayrefs
Alter find_jobs so that it can find jobs for multiple different values
for 'dst_cell' or 'state'. That is, if the caller specifies an
arrayref for 'dst_cell' or 'state', we'll find jobs for any of the
given destination cells and any of the given states.

This also changes find_jobs to return the job state as one of the
returned fields, and lets find_jobs be called without a dbhandle (so
it can be called more easily outside of DB.pm).

Change find_update_jobs to make use of this functionality for
'dst_cell'; future commits will make use of the functionality for
'state'.