Relax --delete requirements through a new option
The original cellcc start-sync --delete flag is used to delete the
specified volume in all destination cells. This option only works if
the volume in question was already removed from the source cell.
Unfortunately, this is not always desirable.

This commit introduces a new flag called --force-delete. This option
deletes the specified volume in all destination cells even if this
volume still exists in the source cell.