Do not create an empty volume before restoration
Currently, the restore daemon checks if the volume received from the dump
server is present in its local cell. If not, the following actions will be
performed by this process:

1. An empty volume with the same name will be created;
2. The replicas will be added;
3. The empty volume will be released;

At this point, the dump file received from the dump server is restored on
top of this empty volume. After that, the volume in question is released.

Unfortunately, this current approach can be problematic. As one can
imagine, the update time of the empty volume is gonna be newer than the
update time of the volume being restored. As a result, the incremental
release of the latter will not update the replicas correctly.

To fix this problem, do not create an empty volume even if the volume in
question does not exist in the destination cell. Instead, let the restore
operation create this volume for us (if necessary).