Do not fail if status of target volume is ok
Currently, CellCC does not dump / restore a volume if the status of this
volume (in at least one server) is equal to one of the following values:
'New release', 'Old release', or 'Not released'. Fortunately, this
conservative approach is not really necessary. First of all, the only
status that should be taken into consideration is the status from the
server where the volume is going to be dumped / restored. Secondly, if
the status of this volume (on this server) is equal to 'New release',
this means that this server has the most recent version of the volume
and we should be able to proceed.

To fix this problem, only take into consideration the status of the
volume from the target server. If this status indicates we have the most
recent version, proceed.