Do not release empty volume before restoration
Currently, the restore daemon checks if the volume received from the
dump server is present in its local cell. If not, the following actions
will be performed by this process:

1. An empty volume with the same name will be created;
2. The replicas will be added;
3. The empty volume will be released;

At this point, the dump file received from the dump server is restored
on top of this empty volume. After that, the volume in question is
released.

Unfortunately, this current approach can be problematic. As one can
imagine, the update time of the empty volume will be newer than the
update time of the volume being restored. As a result, when we go to
release the restored volume with real data, it looks like the empty
volumes are newer than the restored volume, and so vos may only send
some incremental data during the release. The remote replicas will
then appear corrupt or partially empty, since they are missing some
data from the volume dump.

Sometimes, the sanity checks during a vos release will force a full
release, avoiding this issue, but not always.

To fix this problem, do not release the empty volume created during the
RESTORE_WORK state when the volume in question does not exist in the
destination cell.