Retry _START/_DONE/etc states
Currently, our *_START and *_DONE-style states aren't retried on
error. Since those states don't do any real work (all we do is
transition the job to another state, like *_WORK), it was unclear when
they were implemented how they would end up in an error state.

But these jobs can of course end up in an error state if we throw an
error while processing that job before its transitioned to another
state. For all of these states, there isn't much code we run during
that time, but at least one thing we must do is contact the database
to do the transition itself. If our database code throws an error (for
example, we fail to reach the database), we could then flag that job
as an ERROR.

If that happens and the database is reachable immediately afterwards,
the job will go into ERROR state, but will never come out, because
retry_job() will complain that the job is "un-retriable".

To fix this, have states like RELEASE_START be retried by
transitioning them back to the same state. There's no more work that
needs to be done, or a different state to go to, since the job will
just get picked up again and transitioned to the appropriate *_WORK
state or similar.