Reap dead children more frequently
Currently, in the dump and restore servers, we perform work in child
processes managed by Parallel::ForkManager, which reaps dead child
processes at the same time when we fork a new process (when we call
start()). If it takes a long time for us to fork a new child process
to perform work, this means that any dead child processes can linger
around for a long time as zombies. This doesn't cause any immediate
problems, but it looks odd and can be concerning for admins.

To avoid this, change our daemons to explicitly check for and reap
dead child processes on every iteration of our loop that looks for
work to do. Zombies can still exist this way, but they should get
cleaned up every "check-interval" seconds, so they won't linger around
for a long time.