Allow extra arguments for vos release
The commands in the vos command suite are the administrative interface
to the Volume Server and Volume Location Server. As we know, CellCC
takes advantage of these commands to synchronize volumes between AFS
cells. Unfortunately, the system administrator currently cannot
specify which options will be passed to vos.

To improve this, accept extra vos arguments in cellcc.json for the
current queue. In the future we may do this for most/all vos
operations, but for now only do this for 'vos release'.