Specify timeout for database requests
Currently, in certain situations, requests to the database may hang
indefinitely, at least when using mysql. This undesirable behavior can
happen when some specific packets are dropped during the communication
between the requestor and the database. As a result, jobs waiting for
the response of the request in question get stuck in the same state
and have to be restarted manually.

To avoid this problem, add a new configuration directive (options) so
the admin can add / alter the settings to be used when contacting the
database, including the maximum number of seconds a database request
can take before being interrupted. If not specified through this new
directive, set the default timeout to 5 minutes.