RPM packaging support for RHEL 6-8, Fedora 33-34
If someone tries to build our rpm on RHEL/CentOS 8 or Fedora 34 and
the 'perl-generators' package isn't installed, the rpm will be built
without the generated perl dependencies. Add perl-generators to our
BuildRequires to make sure those dependencies are never silently
skipped.

The perl-generators package also exists as a dummy package on RHEL 6-7
(in EPEL), so the same spec file can be used on all of these
platforms. Change some text to make it clear that rpm packaging can be
built for RHEL/CentOS 6-8 and Fedora 33-34. Change the packaging to be
for just 'el' instead of 'el7' (and targets 'rpm7' to 'rpm' et al),
since we're not RHEL7-specific now.